Extract DOM template from jquery.template (#6547)
* Extreact DOM template from jquery.template

* Set default template engine

* Small refactoring

* Revert

* Add const

* Extract templates registration to separate module

* Refactoring

* Refactoring

* Rename module
