+++ testing/tests/DevExpress.ui.widgets/overlay.tests.js
@@ -9,7 +9,7 @@ var $ = require("jquery"),
     domUtils = require("core/utils/dom"),
     resizeCallbacks = require("core/utils/resize_callbacks"),
     devices = require("core/devices"),
-    Template = require("ui/widget/jquery.template"),
+    Template = require("ui/widget/template"),
     Overlay = require("ui/overlay"),
     pointerMock = require("../../helpers/pointerMock.js"),
     eventsEngine = require("events/core/events_engine"),
